Fix rounding error of calculating shortest distance
It's dividing the distance by a million here to make it count less in the score, compared to measures of angles and such. However the order of operations here was:
1. divide coord_t-distance by 1M
2. auto-convert the result to float
In the division in step 1 it was still coord_t (integer) so this ends up rounding down in groups of 1 million units (up to 1mm of actual distance). So the fix is to reverse these operations. First cast to float, then divide by 1 million. This results in a float that doesn't need to be cast again. There will be less rounding errors since floats can represent fractional numbers.

Fixes CURA-8090.

+++ src/PathOrderOptimizer.h
@@ -1,4 +1,4 @@
-//Copyright (c) 2020 Ultimaker B.V.
+//Copyright (c) 2022 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #ifndef PATHORDEROPTIMIZER_H
@@ -453,7 +453,7 @@ protected:
             const coord_t distance = (combing_boundary == nullptr)
                 ? getDirectDistance(here, target_pos)
                 : getCombingDistance(here, target_pos);
-            const float score_distance = (seam_config.type == EZSeamType::SHARPEST_CORNER && seam_config.corner_pref != EZSeamCornerPrefType::Z_SEAM_CORNER_PREF_NONE) ? 0 : distance / 1000000;
+            const float score_distance = (seam_config.type == EZSeamType::SHARPEST_CORNER && seam_config.corner_pref != EZSeamCornerPrefType::Z_SEAM_CORNER_PREF_NONE) ? 0 : static_cast<float>(distance) / 1000000;
             const float corner_angle = (clockwise ? LinearAlg2D::getAngleLeft(previous, here, next) : LinearAlg2D::getAngleLeft(next, here, previous)) / M_PI - 1; //Between -1 and 1.
 
             float score;
